requirements.txt
monai>=1.2.0
torch>=2.0.0
torchvision>=0.15.0
pydicom>=2.4.0
nibabel>=5.1.0
SimpleITK>=2.2.0
numpy>=1.24.0
opencv-python>=4.8.0
matplotlib>=3.7.0
pandas>=2.0.0
scikit-learn>=1.3.0
tensorboard>=2.13.0
apache-airflow>=2.7.0
pyyaml>=6.0.0
tqdm>=4.65.0
albumentations>=1.3.0 